Fix dtype

+++ src/pytesmo/validation_framework/metric_calculators.py
@@ -1259,7 +1259,7 @@ class RollingMetrics(MetadataMetrics):
         window_size_jd = pd.Timedelta(
             window_size).to_numpy()/np.timedelta64(1, 'D')
         pr_arr, rmsd_arr = metrics.rolling_pr_rmsd(
-            timestamps,
+            timestamps.astype('float32'),
             xy[:, 0],
             xy[:, 1],
             window_size_jd,
